--- blocks/signup/constants.js
@@ -1,28 +1,0 @@
-const CONSTANTS = {
-  SIGNUP_HEADING: 'signupHeading',
-  SIGNUP_NEWSLETTER_CHECKBOX: 'signupNewsletterCheckbox',
-  SIGNUP_EALERTS_CHECKBOX: 'signupEalertsCheckbox',
-  SIGNUP_NAME_PLACEHOLDER: 'signupNamePlaceholder',
-  SIGNUP_EMAIL_PLACEHOLDER: 'signupEmailPlaceholder',
-  SIGNUP_BUTTON_TEXT: 'signupButtonText',
-  SIGNUP_TERMS: 'signupTerms',
-  SIGNUP_TERMS_VALIDATION: 'signupTermsValidation',
-  SIGNUP_EMAIL_VALIDATION_MESSAGE: 'signupEmailValidationMessage',
-  SIGNUP_SUCCESS_MESSAGE: 'signupSuccessMessage',
-  SIGNUP_NEWSLETTER_EALERT_ERROR: 'signupNewsletterEalertError',
-  SIGNUP_ERROR_MESSAGE: 'signupErrorMessage',
-  SIGNUP_SUCCESS_MESSAGE_PROFILE_UPDATE: 'signupSuccessMessageProfileUpdate',
-  SIGNUP_ERROR_MESSAGE_EALERT: 'signupErrorMessageEalert',
-  SIGNUP_DESCRIPTION: 'signupDescription',
-  SIGNUP_THANK_YOU_MESSAGE: 'signupSuccessMessageProfileUpdate',
-  SIGNUP_CONFIRMATION_MESSAGE: 'signupSuccessMessage',
-  SIGNUP_CONSENT_API_URL: 'signupConsentApiUrl',
-  SIGNUP_OCP_APIM_SUBSCRIPTION_KEY: 'signupOcpApimSubscriptionKey',
-  SIGNUP_SUBSCRIBE_API_URL: 'signupSubscribeApiUrl',
-  SIGNUP_CUSWBG_SUBSCRIPTION_LIST: 'signupCuswbgSubscriptionList',
-  SIGNUP_EVENT_SUB_LIST: 'signupEventSubList',
-  SIGNUP_EVENT_SUB_UPDATE_LIST: 'signupEventSubUpdateList',
-  SIGNUP_SUBSCRIPTION_TYPE: 'signupSubscriptionType',
-};
-
-export default CONSTANTS;
