+++ blocks/signup/signup.css
@@ -26,7 +26,7 @@
 .email-signup-container .checkbox-group {
   display: flex;
   gap: 8px;
-  margin-top: 10px;
+  margin-top: 18px;
   align-items: start;
   font-size: var(--body-tiny);
 }
@@ -95,11 +95,11 @@
   background-image: url("../../icons/exclamation.png");
   background-repeat: no-repeat;
   padding-left: 14px;
-  
+
 }
 
 
-.email-signup-container input[type="email"].input-error, 
+.email-signup-container input[type="email"].input-error,
 .email-signup-container [type="checkbox"]:not(:checked).input-error {
   border: 1px solid red;
   background-color: #FDF2F2;
@@ -108,6 +108,7 @@
 
 /* Tablet Styles */
 @media (width >=768px) {
+
   .email-signup-container input[type="email"],
   .email-signup-container input[type="text"] {
     padding: 16px;
@@ -155,7 +156,21 @@
 }
 
 @media (width >=1024px) {
+  .email-signup-container {
+    display: flex;
+  }
+
   .email-signup-container .main-heading p {
     margin-bottom: 16px;
   }
+
+  .email-signup-container .signup-content {
+    padding: 0px 60px 0 0px;
+    max-width: 500px;
+  }
+
+  .email-signup-container #signup-form {
+    margin: 20px 0 20px 0;
+  }
+
 }
